requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.29
pydantic==2.7.1
python-multipart==0.0.9
requests==2.32.3
PyYAML==6.0.1
opencv-python-headless==4.9.0.80
pytest==8.1.1
PyJWT==2.8.0
boto3==1.34.131
google-cloud-storage==2.16.0
psycopg2-binary==2.9.9
Jinja2==3.1.4
prometheus-client==0.20.0
kafka-python==2.0.2
paho-mqtt==1.6.1
